Fix operators.

+++ paddle/fluid/operators/math/cross_entropy.cu
@@ -28,7 +28,7 @@ __global__ void CrossEntropyKernel(T* Y, const T* X, const int64_t* label,
   for (int i = blockIdx.x * blockDim.x + threadIdx.x; i < N;
        i += blockDim.x * gridDim.x) {
     PADDLE_ASSERT(label[i] >= 0 && label[i] < D);
-    Y[i] = -math::TolerableValue<T>()(log(X[i * D + label[i]]));
+    Y[i] = -math::TolerableValue<T>()(logf(X[i * D + label[i]]));
   }
 }
